src/main/java/com/ibm/ioc/Annotations.java
/*
* Copyright (c) IBM Corporation 2020. All Rights Reserved.
* Project name: java-ioc
* This project is licensed under the Apache License 2.0, see LICENSE.
*/

package com.ibm.ioc;

import java.lang.annotation.ElementType;
import java.lang.annotation.Retention;
import java.lang.annotation.RetentionPolicy;
import java.lang.annotation.Target;

public interface Annotations {

  /**
   * <i>Strictly optional</i> setter method annotation; provides a hint to the bindings framework that the annotated
   * method must be wired to a bindings parameter.
   */
  @Target(ElementType.METHOD)
  @Retention(RetentionPolicy.RUNTIME)
  @interface RequireParameterBinding {
  }

}


